Create Music within Bracket
- Once Bracket is cached, it will be responsible for determining if Music should be loaded from the URL or not.
- So this moves loading Music from the URL down into Bracket instead of up higher. Lots of refactoring got the code there. whew.
- Move around URL and VaultError to their own files for accessibility to more code.
