+++ Sources/site_tool/RootURLArguments+Lookup.swift
@@ -11,6 +11,6 @@ extension RootURLArguments {
   func lookup<Identifier: ArchiveIdentifier>(identifier: Identifier) async throws
     -> Lookup<Identifier>
   {
-    try await Lookup(music: try await music().showsOnly, identifier: identifier)
+    try await Lookup(url: showsURL, identifier: identifier)
   }
 }
